Move job class into Subscription namespace, shorten description
